Move notification fetching to the main process
