+++ lib/index.js
@@ -12,9 +12,9 @@ import * as uiActions from './actions/ui';
 import * as updaterActions from './actions/updater';
 import * as sessionActions from './actions/sessions';
 import * as termGroupActions from './actions/term-groups';
-import {fetchNotifications} from './actions/notifications';
-import HyperTermContainer from './containers/hyperterm';
+import {addNotificationMessage} from './actions/notifications';
 import {loadConfig, reloadConfig} from './actions/config';
+import HyperTermContainer from './containers/hyperterm';
 import configureStore from './store/configure-store';
 
 // Disable pinch zoom
@@ -37,7 +37,6 @@ config.subscribe(() => {
 // and subscribe to all user intents for example from menus
 rpc.on('ready', () => {
   store_.dispatch(init());
-  store_.dispatch(fetchNotifications());
   store_.dispatch(uiActions.setFontSmoothing());
 });
 
@@ -123,6 +122,10 @@ rpc.on('update available', ({releaseName, releaseNotes}) => {
 
 rpc.on('move', () => {
   store_.dispatch(uiActions.windowMove());
+});
+
+rpc.on('add notification', ({text, url, dismissable}) => {
+  store_.dispatch(addNotificationMessage(text, url, dismissable));
 });
 
 const app = render(
